Rename api-payload module and move package
