+++ api-payload-core/src/main/java/org/namul/api/payload/core/code/BaseSuccessCode.java
@@ -1,0 +1,16 @@
+package org.namul.api.payload.core.code;
+
+import org.springframework.http.HttpStatus;
+
+import java.io.Serializable;
+
+/**
+ * Interface which have base getter to create Success response
+ */
+public interface BaseSuccessCode extends Serializable {
+    /**
+     * The method to get HttpStatus
+     * @return The HttpStatus when success
+     */
+    HttpStatus getHttpStatus();
+}
