+++ api-payload-core/src/main/java/org/namul/api/payload/core/code/supports/DefaultBaseErrorCode.java
@@ -1,0 +1,9 @@
+package org.namul.api.payload.core.code.supports;
+
+import org.namul.api.payload.core.code.BaseErrorCode;
+
+public interface DefaultBaseErrorCode extends BaseErrorCode {
+
+    String getCode();
+    String getMessage();
+}
