+++ api-payload-core/src/main/java/org/namul/api/payload/core/code/supports/DefaultBaseSuccessCode.java
@@ -1,0 +1,9 @@
+package org.namul.api.payload.core.code.supports;
+
+import org.namul.api.payload.core.code.BaseSuccessCode;
+
+public interface DefaultBaseSuccessCode extends BaseSuccessCode {
+
+    String getCode();
+    String getMessage();
+}
