+++ api-payload-core/src/main/java/org/namul/api/payload/core/error/exception/ServerApplicationException.java
@@ -1,0 +1,26 @@
+package org.namul.api.payload.core.error.exception;
+
+import org.namul.api.payload.core.code.BaseErrorCode;
+import lombok.Getter;
+
+/**
+ * The exception that has failure data as BaseErrorCode
+ */
+@Getter
+public class ServerApplicationException extends RuntimeException {
+
+    private final BaseErrorCode code;
+
+    public ServerApplicationException(BaseErrorCode code) {
+        this.code = code;
+    }
+
+    public ServerApplicationException(BaseErrorCode code, Throwable cause) {
+        super(cause);
+        this.code = code;
+    }
+    public ServerApplicationException(BaseErrorCode code, String message, Throwable cause) {
+        super(message, cause);
+        this.code = code;
+    }
+}
