+++ api-payload-core/src/main/java/org/namul/api/payload/core/response/BaseResponse.java
@@ -1,0 +1,9 @@
+package org.namul.api.payload.core.response;
+
+import java.io.Serializable;
+
+/**
+ * The top-level response interface
+ */
+public interface BaseResponse extends Serializable {
+}
