+++ api-payload-core/src/main/java/org/namul/api/payload/core/writer/FailureResponseWriter.java
@@ -1,0 +1,21 @@
+package org.namul.api.payload.core.writer;
+
+import org.namul.api.payload.core.code.BaseErrorCode;
+import org.namul.api.payload.core.response.BaseResponse;
+
+/**
+ * The class that write responses in case of failure.
+ * if you make interface or class with BaseErrorCode, you can implement this interface and make class which create response with your own interface
+ * @param <T> The interface implementing BaseErrorCode which have method to make response
+ */
+public interface FailureResponseWriter<T extends BaseErrorCode> {
+
+    /**
+     * Make failure response.
+     * @param error The data about how to fail
+     * @param t The throwable data
+     * @return The generated return value
+     */
+    BaseResponse onFailure(Throwable t, T error);
+
+}
