+++ api-payload-core/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+org.namul.api.payload.core.config.DefaultResponseWriterAutoConfiguration
