--- api-payload/src/main/java/org/namul/api/payload/code/supports/DefaultBaseErrorCode.java
@@ -1,9 +1,0 @@
-package org.namul.api.payload.code.supports;
-
-import org.namul.api.payload.code.BaseErrorCode;
-
-public interface DefaultBaseErrorCode extends BaseErrorCode {
-
-    String getCode();
-    String getMessage();
-}
