--- api-payload/src/main/java/org/namul/api/payload/code/supports/DefaultBaseSuccessCode.java
@@ -1,9 +1,0 @@
-package org.namul.api.payload.code.supports;
-
-import org.namul.api.payload.code.BaseSuccessCode;
-
-public interface DefaultBaseSuccessCode extends BaseSuccessCode {
-
-    String getCode();
-    String getMessage();
-}
