--- api-payload/src/main/java/org/namul/api/payload/response/BaseResponse.java
@@ -1,9 +1,0 @@
-package org.namul.api.payload.response;
-
-import java.io.Serializable;
-
-/**
- * The top-level response interface
- */
-public interface BaseResponse extends Serializable {
-}
