--- api-payload/src/main/java/org/namul/api/payload/writer/FailureResponseWriter.java
@@ -1,21 +1,0 @@
-package org.namul.api.payload.writer;
-
-import org.namul.api.payload.code.BaseErrorCode;
-import org.namul.api.payload.response.BaseResponse;
-
-/**
- * The class that write responses in case of failure.
- * if you make interface or class with BaseErrorCode, you can implement this interface and make class which create response with your own interface
- * @param <T> The interface implementing BaseErrorCode which have method to make response
- */
-public interface FailureResponseWriter<T extends BaseErrorCode> {
-
-    /**
-     * Make failure response.
-     * @param error The data about how to fail
-     * @param t The throwable data
-     * @return The generated return value
-     */
-    BaseResponse onFailure(Throwable t, T error);
-
-}
