--- api-payload/src/main/java/org/namul/api/payload/writer/supports/DefaultFailureResponseWriter.java
@@ -1,14 +1,0 @@
-package org.namul.api.payload.writer.supports;
-
-import org.namul.api.payload.code.supports.DefaultBaseErrorCode;
-import org.namul.api.payload.response.BaseResponse;
-import org.namul.api.payload.response.supports.DefaultResponse;
-import org.namul.api.payload.writer.FailureResponseWriter;
-
-public class DefaultFailureResponseWriter implements FailureResponseWriter<DefaultBaseErrorCode> {
-
-    @Override
-    public BaseResponse onFailure(Throwable t, DefaultBaseErrorCode error) {
-        return new DefaultResponse<>(false, error.getCode(), error.getMessage(), t == null ? null : t.getMessage());
-    }
-}
